+++ src/scene/mod.rs
@@ -10,6 +10,7 @@
 mod box_handles;
 mod connector;
 pub(crate) mod drag;
+mod frame_request;
 pub(crate) mod node;
 mod scene_inner;
 pub mod toolbar;